+++ src/Hacks/Level/CheckpointFix/CheckpointFix.hpp
@@ -1,0 +1,15 @@
+#pragma once
+
+#include <Module.hpp>
+
+using namespace geode::prelude;
+
+class CheckpointFix : public Module
+{
+    public:
+        MODULE_SETUP(CheckpointFix)
+        {
+            setID("checkpoint-fix");
+            setCategory("Level");
+        }
+};